Allow presentation of dataflow results via rustc errors.
In particular, extended `#[rustc_mir(..)]` so that you can mix in
sub-attributes (namely `dataflow_info_maybe_init` and
`dataflow_info_maybe_uninit`).

If you do one of the above, and (at least for now) also have the
`#[rustc_borrowck_mir]`, then the compiler will emit an error message
for each basic block that includes a line with the name of the
dataflow analysis and the (interpreted) bits that are set after the
dataflow analysis is run on the MIR for the code.

This feature is meant to enable source-code oriented unit testing for
MIR dataflow analogous to e.g. the various variance tests in
`tests/compile-fail/`.